LayoutTests/storage/domstorage/sessionstorage/private-browsing-affects-storage-expected.txt

This is a test to make sure that when private browsing is on any attempt to change the sessionStorage area fail.
Initial value of testItem is: InitialValue
Caught exception trying to change item: Error: QUOTA_EXCEEDED_ERR: DOM Exception 22
After change attempt, testItem is: InitialValue
After remove attempt, testItem is: InitialValue
After clear attempt, testItem is: InitialValue
